% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obst.R
\name{as.list.Gabel}
\alias{as.list.Gabel}
\title{Coerce a Gabel object to a nested list}
\usage{
\method{as.list}{Gabel}(node)
}
\arguments{
\item{node}{(Gabel)}
}
\value{

}
\description{
Coerce a Gabel object to a nested list
}
