% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{ATTENTION!! method predict YET TO BE IMPLEMENTED.
ATTENTION!! method cart YET TO BE IMPLEMENTED.}
\usage{
bagging(B, x_train, x_test, regression = TRUE)
}
\arguments{
\item{B}{integer: number of bootstrap samples}

\item{x_train}{data.frame: training data of dimensions Number of Samples x Features}

\item{x_test}{data.frame: test data of dimensions Number of Samples x Features}

\item{regression}{logical: specify whether tree is a classification or a
regression tree. default=TRUE, TRUE for regression, FALSE for classification}
}
\value{
vector: of size Number of Samples containing bagged predictions to dataset
character vector for classification case and double vector for regression case
}
\description{
Performs bagging on a dataset and returns predictions.
}
