% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{C_hat}
\alias{C_hat}
\title{Minimizer function for classification}
\usage{
C_hat(y1, y2)
}
\arguments{
\item{y1}{(\code{numeric})}

\item{y2}{(\code{numeric})}
}
\description{
Minimizer function for classification
}
