% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Calculate Accuracy}
\usage{
calc_accuracy(ret)
}
\description{
Method that calculated the accuracy the different methods achieved in the iris classification
dataset
}
