% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{calc_mean_accuracy}
\alias{calc_mean_accuracy}
\title{calc_mean_accuracy}
\usage{
calc_mean_accuracy(reps = 100)
}
\arguments{
\item{reps}{rows of data}
}
\description{
Calculate Mean Accuracy of Iris Dataset
}
