% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_greedy}
\alias{cart_greedy}
\title{Create a regression tree greedily based on training data.}
\usage{
cart_greedy(df, depth = 10, threshold = 1, mode = "regression")
}
\arguments{
\item{df}{A data frame containing the training data (X_i1 ... X_id Y_i) as columns.}

\item{steps}{The amount of iterations before halting the algorithm. If unspecified,}
}
\value{

}
\description{
Create a regression tree greedily based on training data.
}
