% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{cart_prune}
\alias{cart_prune}
\title{Prune a CART}
\usage{
cart_prune(node, mask)
}
\arguments{
\item{node}{(\code{Gabel}) node below whitch the mask will be set to FALSE}

\item{mask}{(\code{vector(logical)}) mask for the CART in dfs-order}
}
\value{
mask (\code{vector(logical)}) mask for the pruned CART in dfs-order
}
\description{
returns a mask of the CART with subtree pruned at given node
}
\examples{
mRoot <- cart_prune(Cart$root, mask)
}
