% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forests.R
\name{cart_random_tree}
\alias{cart_random_tree}
\title{Create a regression tree random based on training data.}
\usage{
cart_random_tree(XY, depth = 10L, threshold = 1L, sample = FALSE)
}
\arguments{
\item{XY}{matrix with columns \eqn{1..d} (representing the training data
\eqn{X_i}) and column \eqn{y} (for the values \eqn{Y_i}). Unless sample
is set to TRUE, it is required that there are no observations
\eqn{(X_{i_1}, Y_{i_1})} and \eqn{(X_{i_2}, Y_{i_2})} with \eqn{X_{i_1}
  = X_{i_2}}, but \eqn{Y_{i_1} \neq Y_{i_2}}.}

\item{depth}{The amount of steps before halting the algorithm (defaults to
10)}

\item{threshold}{(integer)}

\item{sample}{(logical)}

\item{mode}{}
}
\value{

}
\description{
Create a regression tree random based on training data.
}
