% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_m}
\alias{compare_m}
\title{Method to qualitatively compare Prediction
Quality of the Random Forests for
different parameters m.}
\usage{
compare_m(m_list, d, n, B)
}
\description{
Method to qualitatively compare Prediction
Quality of the Random Forests for
different parameters m.
}
