% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_m_PE}
\alias{compare_m_PE}
\title{Method to quantitavely compare Prediction
Quality of the Random Forests for
different parameters m.
compare_m_PE(list(1L, 3L, 5L, 10L), d=20, n=1000, B=100L, reps=1)
CAUTION CPU EXPENSIVE TAKES LONG}
\usage{
compare_m_PE(m_list, d, n, B, reps = 400)
}
\description{
Method to quantitavely compare Prediction
Quality of the Random Forests for
different parameters m.
compare_m_PE(list(1L, 3L, 5L, 10L), d=20, n=1000, B=100L, reps=1)
CAUTION CPU EXPENSIVE TAKES LONG
}
