% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_methods_PE}
\alias{compare_methods_PE}
\title{Method to quantitavely compare Prediction
Quality of the four different methods for
high dimensional data.}
\usage{
compare_methods_PE(d, n, B = 100L, reps = 400)
}
\description{
Method to quantitavely compare Prediction
Quality of the four different methods for
high dimensional data.
}
\examples{
compare_methods_PE(d=10, n=1000, B=100L, reps=400) CAUTION!! TAKES VERY LONG CPU EXPENSIVE

}
