% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_performance}
\alias{compare_performance}
\title{Method to test performance of quantiles}
\usage{
compare_performance(n, B, depth, sd, k = 10, random_forest = TRUE, reps = 100)
}
\description{
Method to test performance of quantiles
}
