% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{evalCart}
\alias{evalCart}
\title{Evaluates a decision rule}
\usage{
evalCart(Tree, x)
}
\arguments{
\item{Tree}{the CART to evaluate.}

\item{x}{a numeric vector representing a given data point x.}
}
\value{

}
\description{
The decision rule on a CART is given by:
\deqn{f(x) = \sum_{m=1}{#T} y_m I_{A_{m}(x)}}
where \eqn{x\in\prod([a_i,b_i])} and \eqn{A_m} is the partition induced by the CART.
}
