% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obst.R
\name{Gabel}
\alias{Gabel}
\title{R6 class representing a CART node}
\description{
R6 class representing a CART node

R6 class representing a CART node
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{childL}}{(Gabel)}

\item{\code{childR}}{(Gabel)}

\item{\code{parent}}{(Gabel)}

\item{\code{label}}{unique label for node (integer, defaults to NA_integer_)}

\item{\code{depth}}{(integer)}

\item{\code{points}}{(matrix)}

\item{\code{s}}{(numeric)}

\item{\code{j}}{(integer)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setChildren}{\code{Gabel$setChildren()}}
\item \href{#method-isObst}{\code{Gabel$isObst()}}
\item \href{#method-print}{\code{Gabel$print()}}
\item \href{#method-clone}{\code{Gabel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setChildren"></a>}}
\if{latex}{\out{\hypertarget{method-setChildren}{}}}
\subsection{Method \code{setChildren()}}{
create edges between tree nodes (bi-directional)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gabel$setChildren(Node1, Node2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isObst"></a>}}
\if{latex}{\out{\hypertarget{method-isObst}{}}}
\subsection{Method \code{isObst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gabel$isObst()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of all leaves
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of node attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gabel$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gabel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
