% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_mult_data}
\alias{generate_mult_data}
\title{Generate Data from a multivariate Gaussian}
\usage{
generate_mult_data(n, d, sd = 0.01, mu = NULL, sigma = NULL)
}
\arguments{
\item{n}{dimensional vector with feature means}

\item{d}{dimension}

\item{sd}{standard deviation of irreducible error in y}

\item{mu}{mean for the multivariate normal distribution}

\item{sigma}{positive definite square covariance matrix}
}
\value{
list(df of data, feature means vector, covariance matrix)
}
\description{
Generate Data from a multivariate Gaussian
}
\examples{
generate_mult_data(n=1000, d=5)
}
