% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{isCart.data.tree}
\alias{isCart.data.tree}
\title{Implementation of isCart() for data.tree (data.tree package).}
\usage{
\method{isCart}{data.tree}(Tree)
}
\description{
Implementation of isCart() for data.tree (data.tree package).
}
