% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{isCart}
\alias{isCart}
\title{Check if a tree is a regression or a classification tree.}
\usage{
isCart(Tree)
}
\arguments{
\item{Tree}{the tree to check. isCart() is generic and can be implemented for any graph data structure (e.g. igraph).}
}
\value{
1 for a regression tree, 2 for a classification tree, and 0 otherwise.
}
\description{
This means:
\itemize{
\item The tree is a binary tree.
\item Every inner node (a node with 2 child nodes) has a split index and a split
point attribute.
\item Every leaf has a value y as attribute. For a regression tree, y is numeric;
for a classification tree, y is an integer.
}
}
\seealso{
\code{\link[=isRegressionTree]{isRegressionTree()}}, \code{\link[=isClassificationTree]{isClassificationTree()}}
}
