% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{isClassificationTree}
\alias{isClassificationTree}
\title{Use isCart() to check if a given tree is a classification tree}
\usage{
isClassificationTree(Tree)
}
\arguments{
\item{Tree}{the tree to check.}
}
\value{
TRUE if Tree is a regression tree, FALSE otherwise.
}
\description{
Use isCart() to check if a given tree is a classification tree
}
