% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_iris}
\alias{load_iris}
\title{Loads and returns Iris dataset for classification testing}
\usage{
load_iris()
}
\value{
150 x 5 data.frame
}
\description{
Loads and returns Iris dataset for classification testing
}
