% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{plot_3D}
\alias{plot_3D}
\title{CPU Heavy!!!
Method to 3D render Gaussian Multivariates estimates coming from
generated samples or predicted values}
\usage{
plot_3D(df)
}
\arguments{
\item{df}{data.frame with columns x, y and z. samples are in the rows}
}
\description{
CPU Heavy!!!
Method to 3D render Gaussian Multivariates estimates coming from
generated samples or predicted values
}
