% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{plot_3D_compare_m}
\alias{plot_3D_compare_m}
\title{Method to plot 3D comparisons of prediction SQUARED DIFFERENCES
between ground truth and Random Forest Predictions for
different m}
\usage{
plot_3D_compare_m(path, render = FALSE)
}
\arguments{
\item{path}{character specifying path to compare_RF_m file}

\item{render}{logical should 3D plots be rendered? if TRUE HIGH CPU LOAD!!
if FALSE only 2D plots are returned}
}
\description{
Method to plot 3D comparisons of prediction SQUARED DIFFERENCES
between ground truth and Random Forest Predictions for
different m
}
