% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{plotCart}
\alias{plotCart}
\title{Visualize a regression or classification tree}
\usage{
plotCart(Tree)
}
\arguments{
\item{Tree}{the tree to check.}
}
\description{
The visualisation includes node attributes (compare Richter19, 6.1)
}
