% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{pred_plot_bagging_class}
\alias{pred_plot_bagging_class}
\title{(sin(sqrt(x^2+y^2)))/(sqrt(x^2+y^2))}
\usage{
pred_plot_bagging_class(B, depth, sigma = 0.25, n = 150, random_forest = FALSE)
}
\arguments{
\item{depth}{Integer depths of the bagging generated decision tree}
}
\description{
(sin(sqrt(x^2+y^2)))/(sqrt(x^2+y^2))
}
