% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_greedy}
\alias{pred_plot_greedy}
\title{plots prediction of CART generated regression tree}
\usage{
pred_plot_greedy(
  depth,
  sigma = 0.25,
  n = 150,
  random_forest = FALSE,
  simul = FALSE
)
}
\arguments{
\item{depth}{Integer depths of the CART generated regression tree
pred_plot_greedy(5, sigma=0.25, n=150)}
}
\description{
plots prediction of CART generated regression tree
}
