% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_greedy_class}
\alias{pred_plot_greedy_class}
\title{pred_plot_greedy_class}
\usage{
pred_plot_greedy_class(depth, sigma = 0.25, n = 150, random_forest = FALSE)
}
\arguments{
\item{depth}{The amount of steps before halting the algorithm}

\item{sigma}{standard deviation of irreducible error in y}

\item{n}{number of generated data pairs}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging}
}
\description{
plots prediction of CART generated decision tree
}
\examples{
pred_plot_greedy_class(5, sigma=0.25, n=150)
}
