% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_iris_class}
\alias{pred_plot_iris_class}
\title{Method to Plot the classifcation of iris test dataset}
\usage{
pred_plot_iris_class(depth, B = 100L)
}
\arguments{
\item{depth}{The amount of steps before halting the algorithm}

\item{B}{integer number of bootstrap samples}
}
\description{
Method to Plot the classifcation of iris test dataset
}
