% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_rf}
\alias{pred_plot_rf}
\title{Method plotting one dimension of a multidimensional regression
problem of a multicariate gaussian using a random forest.}
\usage{
pred_plot_rf(n, d, sd, B, depth, m, display_d = 1L, simul = FALSE)
}
\description{
Method plotting one dimension of a multidimensional regression
problem of a multicariate gaussian using a random forest.
}
