% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_sine2D}
\alias{pred_plot_sine2D}
\title{pred_plot_sine2D}
\usage{
pred_plot_sine2D(n, B, depth, sd, k = 10, random_forest = TRUE)
}
\arguments{
\item{n}{number of generated data pairs}

\item{B}{integer number of bootstrap samples}

\item{depth}{The amount of steps before halting the algorithm}

\item{sd}{standard deviation of irreducible error in y}

\item{k}{integer specifying field of view in sine2D case}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging

Method to Plot Predicted Density of a Random Forest
for a two dimensional case for a sine generated dataset
(sin(sqrt(x^2+y^2)))/(sqrt(x^2+y^2))
pred_plot_sine2D(n=1000, B=10L, depth=5, sd=0.1, k=10)}
}
\description{
pred_plot_sine2D
}
