% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_iris}
\alias{prepare_iris}
\title{prepare_iris}
\usage{
prepare_iris(training_set_ratio = 0.8, training_split_indices = NULL)
}
\arguments{
\item{training_set_ratio}{double between 0 and 1 specifies the training test
set ratio}

\item{training_split_indices}{Vector of Integers specifying which rows of the
full set should be taken to be in the training set. Default is NULL. If default
rows are chosen randomly}
}
\value{
Classification Training and Classification Test Set
}
\description{
Method that prepares Iris dataset for classification with Tree-Methods.
setosa == 1, versicolor == 2, virginica == 3
}
