% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{pruneBranch.data.tree}
\alias{pruneBranch.data.tree}
\title{Implementation of pruneBranch() for data.tree}
\usage{
\method{pruneBranch}{data.tree}(Tree, Node)
}
\arguments{
\item{Tree}{the tree to prune.}

\item{Node}{the first node of the branch to remove.}
}
\value{
Deep copy of the the matching nodes of the original tree
(data.tree.Clone)
}
\description{
Implementation of pruneBranch() for data.tree
}
