% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{pruneBranch}
\alias{pruneBranch}
\title{Return a pruned subtree of a given tree,}
\usage{
pruneBranch(Tree, Node)
}
\arguments{
\item{Tree}{the tree to prune.}

\item{Node}{the first node of the branch to remove.}
}
\value{
Subtree with the same root. The return value should be a copy and not
modify the specified tree.
}
\description{
A pruned subtree is a subtree with the same root as the tree it is taken from.
}
