% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{R_hat}
\alias{R_hat}
\title{Minimizer function for regression}
\usage{
R_hat(y1, y2)
}
\arguments{
\item{y1}{(\code{numeric})}

\item{y2}{(\code{numeric})}
}
\description{
Minimizer function for regression
}
