% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{R_hat_n}
\alias{R_hat_n}
\title{Risk factor R_n depending on split point and split index. (regression tree)}
\usage{
R_hat_n(s, j, A, n)
}
\arguments{
\item{s}{split point (numeric)}

\item{j}{split index (integer)}

\item{A}{partition of training data (data.frame)}

\item{n}{size of training data (integer)}
}
\value{
risk factor
}
\description{
Risk factor R_n depending on split point and split index. (regression tree)
}
