% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{R_node}
\alias{R_node}
\title{Optimize split index and split node for leaf nodes v_1 and v_2, based on the
partition of training data in an inner node v. (regression tree)}
\usage{
R_node(A, n, d)
}
\arguments{
\item{A}{partition of training data (X_i, Y_i) (data.frame)}

\item{n}{size of the set of training data (integer)}

\item{d}{dimension of the training data X_i1..X_id}
}
\description{
Ties between found minima are broken at random.
}
