% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rf_plot}
\alias{rf_plot}
\title{Method for generating plots in the shiny app}
\usage{
rf_plot(
  datatype,
  n,
  d = 5,
  m = 1,
  B = 10L,
  depth = 3,
  display_d = 1L,
  sd = 0.1,
  k = 10,
  grid = NULL,
  random_forest = FALSE
)
}
\arguments{
\item{datatype}{choose either gaussian, sine, sine2D, sineclass, or iris
for different datasets}

\item{n}{integer number of samples in the dataset}

\item{d}{integer number of dimension in the generated dataset}

\item{m}{RF parameter}

\item{B}{integer number of independent bootstrap samples}

\item{depth}{integer depth of the CART generated tree}

\item{display_d}{integer dimension displayed in multivariate case gaussian}

\item{sd}{double > 0 inherent noise in the generated data}

\item{k}{integer specifying field of view in sine2D case}

\item{grid}{specify data points on x axis}

\item{random_forest}{TRUE: random forest, FALSE: bagging}
}
\value{
print(plot)
}
\description{
Method for generating plots in the shiny app
}
