% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{start_shiny_app}
\alias{start_shiny_app}
\title{Starts the Shiny App}
\usage{
start_shiny_app()
}
\value{

}
\description{
Datatype choose a datatype (Dropdown menu)
Number_of_obs number of generated data rows (numeric)
Dimension dimension of the record (numeric)
Display_dimension number of shown dimensions (numeric)
m count of random dimensions (numeric)
Sigma standard deviation of irreducible error in y (numeric)
Random_forest TRUE: random forest, FALSE: bagging (logical)
Depth depths of the CART generated regression tree (numeric)
Bootstrap_samples number of bootstrap samples (numeric)
k borders of the sine2D plot (numeric)
Simulate starts the calculation with the given parameters (button)
}
