% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{Bagging on a data set}
\usage{
bagging(
  B,
  x_train,
  x_test,
  depth = 5,
  m = NULL,
  regression = TRUE,
  use_parallel = FALSE,
  random_forest = FALSE
)
}
\arguments{
\item{B}{integer: number of bootstrap samples}

\item{x_train}{data.frame: training data with labels of dimensions Number of Samples x Features + 1}

\item{x_test}{data.frame: test data with labels of dimensions Number of Samples x Features + 1}

\item{m}{Count of random dimensions in random forest. Is set on heuristically motivated values if m is NULL}

\item{regression}{logical: specify whether tree is a classification or a
regression tree. default=TRUE, TRUE for regression, FALSE for classification}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging}

\item{use_parallel:}{Whether to use parallel computation or not}
}
\value{
vector: of size Number of Samples containing bagged predictions to dataset
character vector for classification case and double vector for regression case
}
\description{
Performs bagging on a data set and returns predictions.
}
