% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obst.R
\name{Baum}
\alias{Baum}
\title{R6 class to build a CART}
\description{
A helper class to build a CART based on \code{Gabel} nodes. Nodes are accessible
through a list, for easy introspection with \code{apply}. Nodes are labeled in BFS
order.

As \code{Gabel}, data is stored in environments (\code{R6}, reference semantics).
}
\examples{

## ------------------------------------------------
## Method `Baum$new`
## ------------------------------------------------

T1 <- Baum$new()
T1$root$label # 1L
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nodes}}{List of \code{Gabel} nodes in the tree, in BFS order. A node's
label and its index in this list must match. In particular,
\verb{(Node$label == i) => (identical(nodes[[i]], Node) == TRUE)} holds.}

\item{\code{root}}{The root node of the tree. (\code{Gabel})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Baum$new()}}
\item \href{#method-obstkorb}{\code{Baum$obstkorb()}}
\item \href{#method-depth}{\code{Baum$depth()}}
\item \href{#method-append}{\code{Baum$append()}}
\item \href{#method-validate}{\code{Baum$validate()}}
\item \href{#method-clone}{\code{Baum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Baum object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The root node is initialized with label \code{1L}
and depth \code{0L}.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{T1 <- Baum$new()
T1$root$label # 1L
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obstkorb"></a>}}
\if{latex}{\out{\hypertarget{method-obstkorb}{}}}
\subsection{Method \code{obstkorb()}}{
Return leaf nodes in the tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$obstkorb()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
logical vector with TRUE for leaf nodes, and FALSE otherwise. As
\verb{$nodes}, this vector is in BFS order.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depth"></a>}}
\if{latex}{\out{\hypertarget{method-depth}{}}}
\subsection{Method \code{depth()}}{
Return depth of the tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$depth()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The tree depth is defined as the maximum depth of leaves in the tree.
}

\subsection{Returns}{
(\code{integer})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\if{latex}{\out{\hypertarget{method-append}{}}}
\subsection{Method \code{append()}}{
Add a pair of nodes to the binary tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$append(Parent, Child1, Child2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Parent}}{The parent node to append to. (\code{Gabel})}

\item{\code{Child1}}{The left child node to append. (\code{Gabel})}

\item{\code{Child2}}{The right child node to append. (\code{Gabel})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Check the CART for consistency.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$validate()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The following conditions are verified:
\itemize{
\item all node labels are unique;
\item all leaves have \verb{$j} and \verb{$s} set to \code{NA};
\item all leaves have \verb{$y} not set to \code{NA};
\item all inner nodes have \verb{$j} and \verb{$s} not set to \code{NA};
\item all leaves have at least 1 data point.
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
