% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obst.R
\name{Baum}
\alias{Baum}
\title{R6 class representing a CART}
\description{
R6 class representing a CART

R6 class representing a CART
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Baum$new()}}
\item \href{#method-obstkorb}{\code{Baum$obstkorb()}}
\item \href{#method-append}{\code{Baum$append()}}
\item \href{#method-partition}{\code{Baum$partition()}}
\item \href{#method-predict}{\code{Baum$predict()}}
\item \href{#method-validate}{\code{Baum$validate()}}
\item \href{#method-plot}{\code{Baum$plot()}}
\item \href{#method-clone}{\code{Baum$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Baum object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obstkorb"></a>}}
\if{latex}{\out{\hypertarget{method-obstkorb}{}}}
\subsection{Method \code{obstkorb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$obstkorb()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
logical vector with TRUE for leaf nodes and FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\if{latex}{\out{\hypertarget{method-append}{}}}
\subsection{Method \code{append()}}{
Add a pair of nodes to the binary tree.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$append(label, Child1, Child2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Node}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-partition"></a>}}
\if{latex}{\out{\hypertarget{method-partition}{}}}
\subsection{Method \code{partition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$partition(node, d = nrow(node) - 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{entry point}

\item{\code{d}}{dimension}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$predict(x, node = self$root)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector representing a data point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the expected value by evaluating the tree
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot a CART with 1 or 2-dimensional data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Baum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
