% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{bv_greedy}
\alias{bv_greedy}
\title{bias variance data for 400 reps for CARTs of different depths}
\usage{
bv_greedy(depths_list, sigma = 0.2, n = 150, reps = 400)
}
\description{
bias variance data for 400 reps for CARTs of different depths
}
