% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{bv_plot}
\alias{bv_plot}
\title{bias variance plot for CARTs for different depths}
\usage{
bv_plot(data, plot_title = "Prediction CART Regression Tree", bagging = FALSE)
}
\description{
bias variance plot for CARTs for different depths
}
