% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_greedy}
\alias{cart_greedy}
\title{Create a regression tree greedily based on training data.}
\usage{
cart_greedy(df, depth = 10, threshold = 1, mode = "regression")
}
\arguments{
\item{df}{data frame with columns $1..$d (representing the indices of the
training data X_i) and column $y (representing the corresponding values Y_i)}

\item{steps}{The amount of iterations before halting the algorithm. If unspecified,}
}
\value{

}
\description{
Create a regression tree greedily based on training data.
}
