% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_greedy}
\alias{cart_greedy}
\title{Create a regression tree greedily based on training data.}
\usage{
cart_greedy(XY, depth = 10, threshold = 1, mode = "regression")
}
\arguments{
\item{XY}{matrix with columns $1..$d (representing the
training data X_i) and column $y (for the corresponding values Y_i)}

\item{steps}{The amount of iterations before halting the algorithm
(defaults to 10)}
}
\value{

}
\description{
Create a regression tree greedily based on training data.
}
