% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_greedy}
\alias{cart_greedy}
\title{Create a regression tree greedily based on training data.}
\usage{
cart_greedy(XY, depth = 10L, threshold = 1L, sample = FALSE)
}
\arguments{
\item{XY}{matrix with columns $1..d$ (representing the training data $X_i$)
and column $y$ (for the values $Y_i$). Unless sample is set to TRUE, it is
required that there are no observations $(X_{i_1}, Y_{i_1})$ and $(X_{i_2},
Y_{i_2})$ with $X_{i_1} = X_{i_2}$, but $Y_{i_1} \neq Y_{i_2}$.}

\item{depth}{The amount of steps before halting the algorithm (defaults to
10)}

\item{threshold}{(integer)}

\item{sample}{(logical)}
}
\value{

}
\description{
Create a regression tree greedily based on training data.
}
