% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_greedy}
\alias{cart_greedy}
\title{Create a regression tree greedily based on training data.}
\usage{
cart_greedy(XY, depth = 10L, threshold = 1L, sample = FALSE, random = FALSE)
}
\arguments{
\item{XY}{matrix with columns \eqn{1..d} (representing the training data
\eqn{X_i}) and column \eqn{y} (for the values \eqn{Y_i}). Unless sample
is set to TRUE, it is required that there are no observations
\eqn{(X_{i_1}, Y_{i_1})} and \eqn{(X_{i_2}, Y_{i_2})} with \eqn{X_{i_1}
  = X_{i_2}}, but \eqn{Y_{i_1} \neq Y_{i_2}}.}

\item{depth}{The amount of steps before halting the algorithm (defaults to
10)}

\item{threshold}{(integer)}

\item{sample}{(logical)}

\item{random}{generates CART for random forest (logical)}
}
\value{

}
\description{
Create a regression tree greedily based on training data.
}
