% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{cart_grid}
\alias{cart_grid}
\title{cart_grid}
\usage{
cart_grid(A, d, f, quantile = FALSE, q_threshold = 100L, q_pct = 0.25)
}
\arguments{
\item{A}{data (\code{data.frame} or \code{matrix})}

\item{d}{dimensions (\code{numeric})}

\item{f}{Minimizer \code{R_hat} or \code{C_hat} for regression and classification,
respectively (\code{closure})}

\item{quantile}{use \code{quantile()} instead of data points to compute the risk.
(\code{logical}, defaults to \code{FALSE})}

\item{q_threshold}{minimal amount of data points for using quantiles.
(\code{integer}, defaults to \code{100L})}

\item{q_pct}{amount of probabilities for \code{quantile()}, in pct. of the data
set size, generated with \code{seq()}. (\code{numeric}, defaults to \code{0.25})}
}
\value{
A 3D \code{array} with dimensions (\code{j}, \code{s}, \code{R}).
}
\description{
computes risk R for data points
}
\details{
The data is partitioned with \code{cart_part()}. If \verb{$A1} (the partition
\code{A[, j] < s}) is empty, the corresponding risk is set to \code{NA}
}
\examples{
cart_grid(generate_sin_data(100), 1, R_hat)

}
\seealso{
\link{cart_part}, \link{R_hat}, \link{C_hat}
}
