% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{cart_predict_pruned}
\alias{cart_predict_pruned}
\title{Prediction for pruned CARTs}
\usage{
cart_predict_pruned(x, node, mask)
}
\arguments{
\item{x}{(\code{vector}) one col of XY except y}

\item{node}{(\code{Gabel}) first call with root}

\item{mask}{(\code{vector(logical)}) mask for the CART in dfs-order}
}
\value{
y prediction node$y of appropriate sub-node
}
\description{
returns y of leaf nodes, but considers only parts of the CART not masked
}
