% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_classify_iris}
\alias{compare_classify_iris}
\title{Function that performs classification with all 4 methods on}
\usage{
compare_classify_iris(depth = 5L, B = 100L)
}
\description{
Function that performs classification with all 4 methods on
}
