% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_m}
\alias{compare_m}
\title{compare_m}
\usage{
compare_m(m_list, d, n, B)
}
\arguments{
\item{m_list}{list of parameters m to compare}

\item{d}{dimension}

\item{n}{number of generated data pairs}

\item{B}{number of bootstrap samples

compare_m(m_list = list(1L, 3L, 5L, 10L) ,d=20, n=100, B=100L)}
}
\description{
Method to qualitatively compare Prediction Quality of the Random Forests for different parameters m.
}
