% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_m}
\alias{compare_m}
\title{Method to qualitatively compare Prediction
Quality of the Random Forests for
different parameters m.
compare_m(m_list = list(1L, 3L, 5L, 10L) ,d=20, n=100, B=100L)}
\usage{
compare_m(m_list, d, n, B)
}
\description{
Method to qualitatively compare Prediction
Quality of the Random Forests for
different parameters m.
compare_m(m_list = list(1L, 3L, 5L, 10L) ,d=20, n=100, B=100L)
}
\examples{
compare_m(m_list = list(1L, 3L, 5L, 10L) ,d=20, n=100, B=100L)
}
