% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_m_PE}
\alias{compare_m_PE}
\title{Method to quantitavely compare Prediction
Quality of the Random Forests for
different parameters m.}
\usage{
compare_m_PE(m_list, d, n, reps = 400)
}
\description{
Method to quantitavely compare Prediction
Quality of the Random Forests for
different parameters m.
}
