% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_methods}
\alias{compare_methods}
\title{Method to qualitatively compare Prediction
Quality of the four different methods for
high dimensional data.}
\usage{
compare_methods(d, n, B = 100L)
}
\description{
Method to qualitatively compare Prediction
Quality of the four different methods for
high dimensional data.
}
