% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_methods_PE}
\alias{compare_methods_PE}
\title{Method to quantitavely compare Prediction
Quality of the four different methods for
high dimensional data.
compare_methods_PE(d=4, n=1000L, B=25L, reps=100) CAUTION!!
TAKES VERY LONG CPU EXPENSIVE}
\usage{
compare_methods_PE(d, n, B = 25L, reps = 400)
}
\description{
Method to quantitavely compare Prediction
Quality of the four different methods for
high dimensional data.
compare_methods_PE(d=4, n=1000L, B=25L, reps=100) CAUTION!!
TAKES VERY LONG CPU EXPENSIVE
}
