% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{compare_performance}
\alias{compare_performance}
\title{Method to test performance of quantiles}
\usage{
compare_performance(n, B, depth, sd, k = 10, random_forest = TRUE, reps = 100)
}
\arguments{
\item{n}{number of generated data pairs}

\item{B}{number of bootstrap samples}

\item{depth}{The amount of steps before halting the algorithm}

\item{sd}{standard deviation of irreducible error in y}

\item{k}{specifying field of view in sine2D case}

\item{random_forest}{TRUE: random forest, FALSE: bagging}

\item{reps}{rows of data}
}
\description{
Method to test performance of quantiles
}
