% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_heart_disease}
\alias{load_heart_disease}
\title{Heart Disease Data Set}
\usage{
load_heart_disease()
}
\value{
303 x 14 data.frame with named columnDs
}
\description{
For reference see https://archive.ics.uci.edu/ml/datasets/Heart+Disease
}
\details{
The labels are given by the last column "Diagnosis_Heart_Disease" 0 indicates no heart disease
while 1, 2, 3, 4 indicate a heart disease

data.frame still contains NA and ? values which need to be cleaned before
this can be used for testing and training
}
