% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{plot_3D_compare}
\alias{plot_3D_compare}
\title{Method to plot 3D comparisons of predictions
for different methods}
\usage{
plot_3D_compare(path, render = FALSE, margin = 1)
}
\arguments{
\item{path}{character specifying path to compare_plot_data file}

\item{render}{logical should 3D plots be rendered? if TRUE HIGH CPU LOAD!!
if FALSE only 2D plots are returned}

\item{margin}{margin of the scale

plot_3D_compare("data/simul/compare_plot_data_20200812-113504", render=TRUE)}
}
\description{
Method to plot 3D comparisons of predictions
for different methods
}
