% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{plot_3D_compare_m}
\alias{plot_3D_compare_m}
\title{plot_3D_compare_m}
\usage{
plot_3D_compare_m(path, render = FALSE, margin = 1)
}
\arguments{
\item{path}{character specifying path to compare_RF_m file}

\item{render}{logical should 3D plots be rendered? if TRUE HIGH CPU LOAD!!
if FALSE only 2D plots are returned}

\item{margin}{margin of the scale}
}
\description{
Method to plot 3D comparisons of prediction SQUARED DIFFERENCES
between ground truth and Random Forest Predictions for
different m
}
