% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_bagging}
\alias{pred_plot_bagging}
\title{pred_plot_bagging}
\usage{
pred_plot_bagging(
  depth,
  B,
  sigma = 0.25,
  n = 150,
  grid = NULL,
  random_forest = FALSE,
  simul = FALSE
)
}
\arguments{
\item{depth}{The amount of steps before halting the algorithm}

\item{B}{integer number of bootstrap samples}

\item{sigma}{standard deviation of irreducible error in y}

\item{n}{number of generated data pairs}

\item{grid}{specify data points on x axis}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging}

\item{simul}{True: return data}
}
\description{
plots prediction of Bagging generated regression tree with depth 5
and specified number of bootstrap samples B
}
