% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{pred_plot_bagging}
\alias{pred_plot_bagging}
\title{plots prediction of Bagging generated regression tree with depth 5
and specified number of bootstrap samples B}
\usage{
pred_plot_bagging(
  depth,
  B,
  sigma = 0.25,
  n = 150,
  grid = NULL,
  random_forest = FALSE
)
}
\arguments{
\item{B}{integer number of bootstrap samples}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging}
}
\description{
plots prediction of Bagging generated regression tree with depth 5
and specified number of bootstrap samples B
}
