% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_bagging_class}
\alias{pred_plot_bagging_class}
\title{plots prediction of Bagging generated decision tree}
\usage{
pred_plot_bagging_class(B, depth, sigma = 0.25, n = 150, random_forest = FALSE)
}
\arguments{
\item{depth}{Integer depths of the bagging generated decision tree}
}
\description{
plots prediction of Bagging generated decision tree
}
\examples{
pred_plot_bagging_class(B=10L, depth=5, sigma=0.25, n=150)
}
