% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_bagging_class}
\alias{pred_plot_bagging_class}
\title{plots prediction of Bagging generated decision tree}
\usage{
pred_plot_bagging_class(B, depth, sigma = 0.25, n = 150, random_forest = FALSE)
}
\arguments{
\item{B}{integer number of bootstrap samples}

\item{depth}{The amount of steps before halting the algorithm}

\item{sigma}{standard deviation of irreducible error in y}

\item{n}{number of generated data pairs}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging}
}
\description{
plots prediction of Bagging generated decision tree
}
