% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_greedy}
\alias{pred_plot_greedy}
\title{plots prediction of CART generated regression tree}
\usage{
pred_plot_greedy(
  depth,
  sigma = 0.25,
  n = 150,
  random_forest = FALSE,
  simul = FALSE
)
}
\arguments{
\item{depth}{The amount of steps before halting the algorithm}

\item{sigma}{standard deviation of irreducible error in y}

\item{n}{number of generated data pairs}

\item{random_forest}{logical: TRUE: random forest, FALSE: bagging
pred_plot_greedy(5, sigma=0.25, n=150)}

\item{simul}{True: return data}
}
\description{
plots prediction of CART generated regression tree
}
