% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_greedy_class}
\alias{pred_plot_greedy_class}
\title{plots prediction of CART generated decision tree}
\usage{
pred_plot_greedy_class(depth, sigma = 0.25, n = 150, random_forest = FALSE)
}
\arguments{
\item{depth}{Integer depths of the CART generated decision tree
pred_plot_greedy_class(5, sigma=0.25, n=150)}
}
\description{
plots prediction of CART generated decision tree
}
