% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_rf}
\alias{pred_plot_rf}
\title{pred_plot_rf}
\usage{
pred_plot_rf(n, d, sd, B, depth, m, display_d = 1L, simul = FALSE)
}
\arguments{
\item{n}{number of generated data pairs}

\item{d}{dimension}

\item{sd}{standard deviation of irreducible error in y}

\item{B}{integer number of bootstrap samples}

\item{depth}{The amount of steps before halting the algorithm}

\item{m}{RF parameter: count of chosen dimension}

\item{display_d}{integer dimension displayed in multivariate case gaussian}

\item{simul}{True: return data}
}
\description{
Method plotting one dimension of a multidimensional regression
problem of a multicariate gaussian using a random forest.
}
