% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_plot_sine2D}
\alias{pred_plot_sine2D}
\title{Method to Plot Predicted Density of a Random Forest
for a two dimensional case for a sine generated dataset
(sin(sqrt(x^2+y^2)))/(sqrt(x^2+y^2))}
\usage{
pred_plot_sine2D(n, B, depth, sd, k = 10, random_forest = TRUE)
}
\description{
Method to Plot Predicted Density of a Random Forest
for a two dimensional case for a sine generated dataset
(sin(sqrt(x^2+y^2)))/(sqrt(x^2+y^2))
}
