% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{R_min}
\alias{R_min}
\title{Find optimal split index $j$ and split point $s$ for a given partition A
(regression tree)}
\usage{
R_min(A, d)
}
\arguments{
\item{A}{subset of training data $(X_i, Y_i)$ (matrix)}

\item{d}{dimension of the training data $X_{i_1}..X_{i_d}$}
}
\value{
list containing optimal parameters $j$, $s$
}
\description{
Find optimal split index $j$ and split point $s$ for a given partition A
(regression tree)
}
