% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{start_shiny_app}
\alias{start_shiny_app}
\title{Starts the Shiny App}
\usage{
start_shiny_app()
}
\arguments{
\item{Number_of_obs}{number of generated data rows (numeric)}

\item{Sigma}{standard deviation of irreducible error in y (numeric)}

\item{Random_forest}{TRUE: random forest, FALSE: bagging (logical)}

\item{Depth}{depths of the CART generated regression tree (numeric)}

\item{Bootstrap_samples}{number of bootstrap samples (numeric)}

\item{Simulate}{starts the calculation with the given parameters (button)}
}
\value{

}
\description{
All arguments are selectable in the App
}
